Filter out maximum value in _getAverageFrameDuration
When every value was taken into account, the average value could be
easily skewed by switching to another tab or blocking of the browser
process. This was mostly visible when rewinding the finished playback
to the beginning after it ended while being on another browser tab.
The enormous spike of last frame duration would cause the playback
to jump instantly deep into the playback right after seeking to its
very begining.